Revert "sched/fair: use min capacity when evaluating idle backup cpus"
This reverts commit 88a968c.

Given the experimental nature of this patch and considering the
intention to simplify find_best_target, it's best to not introduce
unnecessary complexity in the heuristics here until the necessity of
this functionality is proven.

This patch could also potentially introduce a preference towards big CPUs
for !prefer_idle and boosted tasks, when iterating though CPUs from bigs
to LITTLES.
Despite it not being a typical schedtune configuration, this contradicts
the intention of selecting energy efficient CPUs for !prefer_idle tasks.

Change-Id: I60217c5d8624b375582946b0c4b3660c0517e3f3
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>
Reported-by: Viresh Kumar <viresh.kumar@linaro.org>
Reported-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
[@0ctobot: This reapplies commit 1d16103]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Joshua Primero <jprimero155@gmail.com>

 Conflicts:
	kernel/sched/fair.c